marines2 update
marines2 has been broken down into configs by camo type.
