+++ src/Addons/34thPRC_ArmourStandard/data/marines2/arctic/config_weapons.hpp
@@ -1,0 +1,380 @@
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic : 34thPRC_ArmourStandard_Marines2_CH252_Urban
+{
+	displayName="[34th] CH252 Arctic";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+
+};class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+
+};class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
+{
+	displayName="[34th] CH252 Arctic (Medic)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Medic) (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Medic) (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Medic) (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Medic) (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arctic (Medic) (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Medic) (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Medic) (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Medic) (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Medic) (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arctic (Medic) (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Light)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Medium)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Heavy)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Shoulderless)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_NS";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Light) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Medium) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Arctic (Heavy) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy_NoCollar";
+	};
+};
