+++ src/Addons/34thPRC_ArmourStandard/data/marines2/arid/config_weapons.hpp
@@ -1,0 +1,372 @@
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid : 34thPRC_ArmourStandard_Marines2_CH252_Urban
+{
+	displayName="[34th] CH252 Arid";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arid (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arid (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arid (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arid (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Arid (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
+{
+	displayName="[34th] CH252 Arid (Medic)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arid (Medic) (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arid (Medic) (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arid (Medic) (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arid (Medic) (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Arid (Medic) (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Medic) (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Medic) (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Medic) (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Medic) (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Arid (Medic) (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Light)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Light";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Medium)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Medium";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Heavy)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Heavy";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Shoulderless)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_NS";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Light) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Light_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Medium) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Medium_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Arid (Heavy) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Heavy_NoCollar";
+	};
+};
