+++ src/Addons/34thPRC_ArmourStandard/data/marines2/config_weapons.hpp
@@ -1,1626 +1,6 @@
 class CH252_Helmet_Standard; // MA_Armor
 class CH252_Helmet_EyePiece; // MA_Armor
 class CH252_Helmet_VisorGlasses; // MA_Armor
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban : CH252_Helmet_Standard
-{
-	displayName="[34th] CH252 Urban";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece : CH252_Helmet_EyePiece
-{
-	displayName="[34th] CH252 Urban (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Urban (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Urban (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-
-};class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Urban (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-
-};class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Urban (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses : CH252_Helmet_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic : 34thPRC_ArmourStandard_Marines2_CH252_Urban
-{
-	displayName="[34th] CH252 Arctic";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-
-};class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-
-};class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid : 34thPRC_ArmourStandard_Marines2_CH252_Urban
-{
-	displayName="[34th] CH252 Arid";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arid (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arid (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arid (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arid (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Arid (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic : 34thPRC_ArmourStandard_Marines2_CH252_Urban
-{
-	displayName="[34th] CH252 Tropic";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland : 34thPRC_ArmourStandard_Marines2_CH252_Urban
-{
-	displayName="[34th] CH252 Woodland";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic : CH252_Helmet_Standard
-{
-	displayName="[34th] CH252 Urban (Medic)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece : CH252_Helmet_EyePiece
-{
-	displayName="[34th] CH252 Urban (Medic) (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Blue : CH252_Helmet_EyePiece
-{
-	displayName="[34th] CH252 Urban (Medic) (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Orange : CH252_Helmet_EyePiece
-{
-	displayName="[34th] CH252 Urban (Medic) (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Red : CH252_Helmet_EyePiece
-{
-	displayName="[34th] CH252 Urban (Medic) (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Purple : CH252_Helmet_EyePiece
-{
-	displayName="[34th] CH252 Urban (Medic) (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses : CH252_Helmet_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Medic) (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Blue : CH252_Helmet_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Medic) (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Green : CH252_Helmet_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Medic) (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Purple : CH252_Helmet_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Medic) (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Red : CH252_Helmet_VisorGlasses
-{
-	displayName="[34th] CH252 Urban (Medic) (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\urban\Medic_Base_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
-{
-	displayName="[34th] CH252 Arctic (Medic)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Medic) (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Medic) (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Medic) (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Medic) (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arctic (Medic) (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Medic) (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Medic) (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Medic) (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Medic) (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arctic (Medic) (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arctic\Medic_White_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
-{
-	displayName="[34th] CH252 Arid (Medic)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arid (Medic) (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arid (Medic) (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arid (Medic) (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arid (Medic) (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Arid (Medic) (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Medic) (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Medic) (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Medic) (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Medic) (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Arid (Medic) (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\arid\Medic_Tan_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
-{
-	displayName="[34th] CH252 Tropic (Medic)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Medic) (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Medic) (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Medic) (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Medic) (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Tropic (Medic) (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Medic) (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Medic) (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Medic) (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Medic) (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Tropic (Medic) (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
-{
-	displayName="[34th] CH252 Woodland (Medic)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
-		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Medic) (Green Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Medic) (Blue Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Medic) (Orange Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Medic) (Red Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
-{
-	displayName="[34th] CH252 Woodland (Medic) (Purple Eye Piece)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Medic) (Orange Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
-	};
-};
-
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Medic) (Blue Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Medic) (Green Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Medic) (Purple Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
-{
-	displayName="[34th] CH252 Woodland (Medic) (Red Visor Glasses)";
-	author="Harrison";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
-		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
-	};
-};
-
 class U_B_CombatUniform_mcam; // A3_Characters_F
 class MA_Marine_BDU_Light: U_B_CombatUniform_mcam // MA_Armor
 {
@@ -1651,381 +31,8 @@ class MA_Marine_BDU_NoShoulder: U_B_CombatUniform_mcam //MA_Armor
 	class ItemInfo;
 };
 
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light: MA_Marine_BDU_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Light)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium: MA_Marine_BDU_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Medium)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy: MA_Marine_BDU_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Heavy)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder: MA_Marine_BDU_NoShoulder
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Shoulderless)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_NS";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar: MA_Marine_BDU_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Light) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar: MA_Marine_BDU_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Medium) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Medium_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar: MA_Marine_BDU_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Red (Heavy) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light: MA_Marine_BDU_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Light)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium: MA_Marine_BDU_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Medium)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy: MA_Marine_BDU_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Heavy)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_NoShoulder: MA_Marine_BDU_NoShoulder
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Shoulderless)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_NS";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light_NoCollar: MA_Marine_BDU_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Light) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium_NoCollar: MA_Marine_BDU_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Medium) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy_NoCollar: MA_Marine_BDU_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Urban Blue (Heavy) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Light)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Medium)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Heavy)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Shoulderless)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_NS";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Light) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Medium) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Medium_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Arctic (Heavy) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Heavy_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Light)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Light";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Medium)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Medium";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Heavy)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Heavy";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Shoulderless)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_NS";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Light) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Light_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Medium) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Medium_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Arid (Heavy) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Arid_Heavy_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Light)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Medium)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Heavy)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Shoulderless)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_NS";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Light) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Medium) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Tropic (Heavy) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Light)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Medium)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Heavy)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Shoulderless)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_NS";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Light) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Medium) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium_NoCollar";
-	};
-};
-class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
-{
-	displayName="[34th] Marine Cross Branch BDU Woodland (Heavy) (No Collar)";
-	author="Ambrose";
-	class ItemInfo: ItemInfo
-	{
-		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy_NoCollar";
-	};
-};
+#include "urban\config_weapons.hpp"
+#include "arid\config_weapons.hpp"
+#include "arctic\config_weapons.hpp"
+#include "tropic\config_weapons.hpp"
+#include "woodland\config_weapons.hpp"
