+++ src/Addons/34thPRC_ArmourStandard/data/marines2/tropic/config_weapons.hpp
@@ -1,0 +1,373 @@
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic : 34thPRC_ArmourStandard_Marines2_CH252_Urban
+{
+	displayName="[34th] CH252 Tropic";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
+{
+	displayName="[34th] CH252 Tropic (Medic)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Medic) (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Medic) (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Medic) (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Medic) (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Tropic (Medic) (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Medic) (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Medic) (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Medic) (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Medic) (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Tropic (Medic) (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\tropic\Medic_Brown_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Light)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Medium)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Heavy)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Shoulderless)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_NS";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Light) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Light_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Medium) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Medium_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Tropic (Heavy) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Tropic_Heavy_NoCollar";
+	};
+};
