+++ src/Addons/34thPRC_ArmourStandard/data/marines2/woodland/config_weapons.hpp
@@ -1,0 +1,389 @@
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland : 34thPRC_ArmourStandard_Marines2_CH252_Urban
+{
+	displayName="[34th] CH252 Woodland";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+
+
+
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
+{
+	displayName="[34th] CH252 Woodland (Medic)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa",
+		"MA_Armor\data\Helmets\CH252\data\MA_HelmetAccessories_A_CO.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Medic) (Green Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Medic) (Blue Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Orange : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Medic) (Orange Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Medic) (Red Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	displayName="[34th] CH252 Woodland (Medic) (Purple Eye Piece)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Medic) (Orange Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Orange_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Blue : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Medic) (Blue Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Blue_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Green : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Medic) (Green Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Green_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Purple : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Medic) (Purple Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Purple_CA.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Red : 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	displayName="[34th] CH252 Woodland (Medic) (Red Visor Glasses)";
+	author="Harrison";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\woodland\Medic_Green_TrooperHelmet.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa",
+		"34thPRC_ArmourStandard\data\marines2\visor\HelmetAccessories_Red_CA.paa"
+	};
+};
+
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Light)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Medium)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Heavy)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NoShoulder: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NoShoulder
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Shoulderless)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_NS";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Light) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Light_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Medium) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Medium_NoCollar";
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy_NoCollar: 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
+{
+	displayName="[34th] Marine Cross Branch BDU Woodland (Heavy) (No Collar)";
+	author="Ambrose";
+	class ItemInfo: ItemInfo
+	{
+		uniformClass="34thPRC_ArmourStandard_Marines2_Trooper_Woodland_Heavy_NoCollar";
+	};
+};
